Breaking things into SFC's
Trying to break apart the components into more manageable chunks
